% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkLog}
\alias{checkLog}
\title{Check if the data needs log-transformation}
\usage{
checkLog(intensities)
}
\arguments{
\item{intensities}{A numeric matrix of gene intensities.}
}
\value{
TRUE or FALSE of whether to proceed with log-transform
}
\description{
Check if the data needs log-transformation
}
