% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_multiple_networks}
\alias{compare_multiple_networks}
\title{Network rewiring function}
\usage{
compare_multiple_networks(x, top_mes)
}
\arguments{
\item{x}{The edge lists as obtained from \code{compute_rewiring}}

\item{top_mes}{Union lists as obtained from \code{compute_rewiring}}
}
\value{
A numeric matrix of the same dimensions with or without normalization/log-transform.
}
\description{
Network rewiring function
}
