% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_rewiring}
\alias{compute_rewiring}
\title{Compute gene rewiring between time points}
\usage{
compute_rewiring(object, ...)
}
\arguments{
\item{object}{A \code{gene_network} object.}
}
\description{
Compute gene rewiring between time points
}
