% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cureNegative}
\alias{cureNegative}
\title{Handle negatives}
\usage{
cureNegative(intensities)
}
\arguments{
\item{intensities}{A numeric matrix of gene intensities.}
}
\value{
A numeric matrix or vector of the same dimensions with all values shifted to be positive.
}
\description{
Handle negatives
}
