% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handleData}
\alias{handleData}
\title{Normalize and/or log-transform expression data}
\usage{
handleData(intensities, norm = FALSE, log = FALSE)
}
\arguments{
\item{intensities}{A numeric matrix of gene intensities.}

\item{norm}{Whether to normalize the data or not.}

\item{log}{Whether to log-transform the data or not.}
}
\value{
A numeric matrix of the same dimensions with ot without normalization/log-transform.
}
\description{
Normalize and/or log-transform expression data
}
