% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{network_construction}
\alias{network_construction}
\title{Create gene co-expression networks from selected intensities}
\usage{
network_construction(object)
}
\arguments{
\item{object}{A \code{gene_network} object.}
}
\description{
Create gene co-expression networks from selected intensities
}
