% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_duplicate_genes}
\alias{remove_duplicate_genes}
\title{Remove duplicate genes (general)}
\usage{
remove_duplicate_genes(eset, column_of_symbol = "GeneSymbol", method = "mean")
}
\arguments{
\item{eset}{A numeric matrix of gene intensities.}

\item{column_of_symbol}{The columns name of gene symbols}
}
\value{
A numeric matrix of the same dimensions without duplicate acquisitions.
}
\description{
Remove duplicate genes (general)
}
