% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_group_intensities}
\alias{select_group_intensities}
\title{Extract intensities for a group, matched across time points}
\usage{
select_group_intensities(object, ...)
}
\arguments{
\item{object}{A \code{gene_network} object.}
}
\description{
Extract intensities for a group, matched across time points
}
