% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_rewiring,gene_network-method}
\alias{compute_rewiring,gene_network-method}
\title{Network rewiring}
\usage{
\S4method{compute_rewiring}{gene_network}(object, edgelist_path, filenames)
}
\arguments{
\item{object}{A \code{gene_network} object.}

\item{edgelist_path}{Path to the directory containing edge list files}

\item{filenames}{Filenames refering to the pathway names}
}
\value{
Export the rewiring scores for each pathway.
}
\description{
Computes network rewiring for a given group and two timepoints.
}
