% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_rewiring}
\alias{compute_rewiring}
\title{Generic for compute_rewiring}
\usage{
compute_rewiring(object, ...)
}
\arguments{
\item{object}{An object}
}
\description{
Generic for compute_rewiring
}
