% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_statistics}
\alias{compute_statistics}
\title{Compute summary statistics from rewiring scores}
\usage{
compute_statistics(scores, filename, dataset, output.folder, pathways)
}
\arguments{
\item{scores}{The rewiring scores}

\item{filename}{The filename corresponding to the group and timepoint label).}

\item{dataset}{The dataset name}

\item{output.folder}{Path to output directory}

\item{pathways}{The pathway names}
}
\value{
Exports the statistical summary for the given dataset and group \code{output.folder}.
}
\description{
Compute summary statistics from rewiring scores
}
