% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{gene_network-class}
\alias{gene_network-class}
\title{Gene co-expression network class}
\description{
A class to represent a pathway-specific gene co-expression network for estimating network rewiring in two time points.
}
\section{Slots}{

\describe{
\item{\code{group_name}}{Name of the group to analyze (e.g. HC, Patients)}

\item{\code{timepoints}}{Vector of timepoints to analyze (numeric or character)}

\item{\code{gene_intensities}}{Matrix of gene expression data, where rows are genes and columns are sample or subject IDs}

\item{\code{metadata}}{Data frame of sample metadata describing the samples, including group, time, subject IDs}

\item{\code{pathways}}{Data frame of pathway/gene sets}

\item{\code{output_dir}}{Path to the output directory}

\item{\code{dataset}}{The name or accession ID of the dataset}

\item{\code{database}}{Name of database (KEGG or GO)}

\item{\code{group_col}}{Column name in `metadata` representing group}

\item{\code{subjects_ids}}{Column name in `metadata` representing subjects IDs}

\item{\code{sample_ids}}{Column name in `metadata` representing sample IDs}

\item{\code{time_col}}{Column name in `metadata` representing timepoints}

\item{\code{time_labels}}{Optional labels for timepoints to use in file names}
}}

