% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{network_construction,gene_network-method}
\alias{network_construction,gene_network-method}
\title{Network construction}
\usage{
\S4method{network_construction}{gene_network}(object)
}
\arguments{
\item{object}{A \code{gene_network} object.}
}
\description{
Constructs adjacency networks for a given group and two timepoints.
}
