% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{network_construction}
\alias{network_construction}
\title{Generic for network_construction}
\usage{
network_construction(object)
}
\arguments{
\item{object}{An object}
}
\description{
Generic for network_construction
}
