% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_group_intensities,gene_network-method}
\alias{select_group_intensities,gene_network-method}
\title{Select group intensities}
\usage{
\S4method{select_group_intensities}{gene_network}(object)
}
\arguments{
\item{object}{A \code{gene_network} object.}
}
\value{
A list with two matrices: gene intensities for time_1 and time_2.
}
\description{
Extract gene intensities for a given group across two timepoints.
}
