% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_group_intensities}
\alias{select_group_intensities}
\title{Generic for select_group_intensities}
\usage{
select_group_intensities(object, ...)
}
\arguments{
\item{object}{An object}
}
\description{
Generic for select_group_intensities
}
